adds new opcode type to account for zmachine instruction strangeness
- so, this was not made clear at all in the documentation ( it is worded
strangely ), but basically, any 20P code can also be encoded in "var"
format.

the documentation makes it sounds like the forms - long, short, and var
- more or less correspond strongly to argument amounts. long is always
2, short 1 or 0, and var 2 or 0.

but what it doesn't make clear is the last part is actually hiding
something else - that 2-OP opcodes can actually be encoded as variable
codes to overcome the limitations of 2-OP codes. the most common of this
is that some 2-OP codes were modified to be able to take large constants
, such as add, which is a current limitation of the 2OP code in long
form.

the CRAZIEST of which is that one 2-OP code , je, can ACTUALLY ( yes,
this deserves a wtf ) take up to 4 operands. So you have a 2-OP code
that...well, lets just stop there.

So now, we have LongVariable, which indicates that Long has been encoded
in variable format, and it decodes its variables just like a variable
format does, but it calls LONG form instructions ( which will then have
to deal with the extra operands if they can ).