adds print_obj, starts input loop, implements rtrue/rfalse on branch
- breaks up zmachine a bit as well
- adds form_base_op_code and form_rfalse, form_rtrue
- adds read_u32_at_head to memory view - we may also need write
- rust fmt all over the place

i may change this pattern to a provider rather than callback oriented -
right now sread sets a callback in the waiting for input struct, but its
possible that instead we want to provide an input handler to zmachine on
init and allow it to block the instruction ( this is probably how it was
traditionally done )

the closure pattern has some negative implications re: changing memory
view's write functions over to mut, if that ever becomes desirable.

im starting to wonder if it makes sense though, given that the Rc itself
can still be cloned as a reference, and then that could be made mutable,
and still write to the same shared memory ( which is the danger behind
Rc<RefCell<>> ).

there may be a better memory pattern to use if we break up the machine
into parts, or just one in general that we failed to see at the
beginning. its entirely possible, for instance, to treat static and high
memory as immutable, and that alone might free up a lot of the design -
SOME implementations actually use the slots in-memory in routines for
local variables, and i was thinking that may have been necessary, but as
it turns out, its actually more complicated and completely unnecessary