major modifications for compatibility with emscripten
- as it turns out, emscripten has to function differently from how a
traditional Rust program might operate with an infinite loop; emscripten
expects you to set a main loop function from main() and then that
function becomes your "new" main.

there are a couple challenges with this; first, that function will be
called every 1/nth of a second, and cannot block ( javascript is single
threaded - this will result in browser lock ).

as a result, you are either a) querying objects and constructing them
every loop or b) using static muts to represent some global, mutable
state. both are gross. a) is really slow in JS... so we choose b).

there are also some unsafe blocks that had to be added because
interacting with enscripten is technically unsafe ( C++ lib ),
and, just to top off a trifecta of things we might have regrets about,
we also had to do some type shadowing to get the desktop version to
compile without installing webplatform.

- its clear there could be a library and a interface implementation
here; not sure if i want to go down that route yet.

- adds a webreader that has a Rc<RefCell<WebInputIndicator>>, which is
just a small struct that tells Rust something happened ( input was
entered ). we do this so we don't have to "block" to wait for input in
emscripten, and then when the flag is set, we pull from the appropriate
html value

we had to do this because part of WebReader has a lifetime - we COULD
achieve this with a static lifetime bound set on WebReader's type, but
the issue is we might actually have to grab a different html element ,
or want that to be an option during runtime, and as a result we have no
guarantees about the lifetime of the callback relating to the webreader.
we can, however, put a struct on the heap, have a few things jointly own
it, and everything will be fine in either scenario.