an attempt at putting the GUI behind the interface
I'm not really sure the ZMachine /OpCode should be generic, I'm starting
to think that they should be using a trait object instead.

This was a rough refactor. In addition I wasn't able to remove as many
flags as I had hoped and ultimately I found that I had to replicate the
main loop for both platforms; I want to get away from this in the future